Module pallet_contracts_mock_network::parachain

Parachain runtime mock.

Structs

ApprovalDeposit

AssetAccountDeposit

AssetDeposit

AssetsStringLimit

BlockHashCount

CheckingAccount

ExistentialDeposit

ForeignPrefix

FromMultiLocationToAsset

KsmLocation

MaxAssetsIntoHolding

MaxInstructions

MaxLocks

MaxReserves

MetadataDepositBase

MetadataDepositPerByte

NftCollectionOne

NftCollectionOneForRelay

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParentRelay

RelayNativeAsset

RelayNativeAssetForRelay

RelayNetwork

RemoveItemsLimit

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

ThisParachain

TokenLocation

TokensPerSecondPerMegabyte

TrustedLockPairs

TrustedLockerCase

UniversalLocation

XcmConfig

XcmInstructionWeight

Enums

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Functions

estimate_fee_for_weight

estimate_message_fee

estimate_weight

Type Definitions

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AssetTransactors

Means for transacting assets on this chain

Assets

AssetsConfig

Balances

BalancesConfig

Barrier

Contracts

ForeignAssetsTransactor

GenesisConfigDeprecated

LocalBalancesTransactor

LocalOriginToLocation

MsgQueue

PolkadotXcm

PolkadotXcmConfig

SovereignAccountOf

System

SystemConfig

Timestamp

TrustedReserves

TrustedTeleporters

XcmOriginToCallOrigin

XcmRouter