pallet_contracts_mock_network

Module relay_chain

Relay chain runtime mock.

Structs

ChildrenParachains

ExistentialDeposit

MaxAssetsIntoHolding

MaxInstructions

MaxLocks

MaxReserves

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RelayNetwork

Runtime

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

TokenLocation

TokensPerSecondPerMegabyte

UnitWeightCost

UniversalLocation

XcmConfig

XcmInstructionWeight

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Type Aliases

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AssetTransactors

Balances

BalancesConfig

Barrier

LocalBalancesTransactor

LocalOriginToLocation

MessageQueue

ParasOrigin

SovereignAccountOf

System

SystemConfig

XcmPallet

XcmPalletConfig

XcmRouter