Module pallet_contracts_mock_network::relay_chain

Relay chain runtime mock.

Structs

BlockHashCount

ChildrenParachains

ExistentialDeposit

MaxAssetsIntoHolding

MaxInstructions

MaxLocks

MaxReserves

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RelayNetwork

Runtime

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

TokenLocation

TokensPerSecondPerMegabyte

UnitWeightCost

UniversalLocation

XcmConfig

XcmInstructionWeight

Enums

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Type Definitions

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AssetTransactors

Balances

BalancesConfig

Barrier

GenesisConfigDeprecated

LocalBalancesTransactor

LocalOriginToLocation

MessageQueue

ParasOrigin

SovereignAccountOf

System

SystemConfig

XcmPallet

XcmRouter